requirements.txt
openai==1.7.0
python-dotenv
gradio==4.14.0
streamlit==1.30.0
colorlog==6.8.0